Restructure application for organization
